SELinux: Institute file_path_has_perm()
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author David Howells <dhowells@redhat.com>
commit 13f8e9810bff12d01807b6f92329111f45218235

Create a file_path_has_perm() function that is like path_has_perm() but
instead takes a file struct that is the source of both the path and the
inode (rather than getting the inode from the dentry in the path).  This
is then used where appropriate.

This will be useful for situations like unionmount where it will be
possible to have an apparently-negative dentry (eg. a fallthrough) that is
open with the file struct pointing to an inode on the lower fs.

	Signed-off-by: David Howells <dhowells@redhat.com>
	Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
(cherry picked from commit 13f8e9810bff12d01807b6f92329111f45218235)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>