bridge: use the bridge IP addr as source addr for querier
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Cong Wang <amwang@redhat.com>
commit 1c8ad5bfa2be5025b0c81e3c2decd0574d453ab1

Quote from Adam:
"If it is believed that the use of 0.0.0.0
as the IP address is what is causing strange behaviour on other devices
then is there a good reason that a bridge rather than a router shouldn't
be the active querier? If not then using the bridge IP address and
having the querier enabled by default may be a reasonable solution
(provided that our querier obeys the election rules and shuts up if it
sees a query from a lower IP address that isn't 0.0.0.0). Just because a
device is the elected querier for IGMP doesn't appear to mean it is
required to perform any other routing functions."

And introduce a new troggle for it, as suggested by Herbert.

	Suggested-by: Adam Baker <linux@baker-net.org.uk>
	Cc: Herbert Xu <herbert@gondor.apana.org.au>
	Cc: Stephen Hemminger <stephen@networkplumber.org>
	Cc: "David S. Miller" <davem@davemloft.net>
	Cc: Adam Baker <linux@baker-net.org.uk>
	Signed-off-by: Cong Wang <amwang@redhat.com>
	Acked-by: Herbert Xu <herbert@gondor.apana.org.au>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 1c8ad5bfa2be5025b0c81e3c2decd0574d453ab1)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>