bridge: multicast: start querier timer when running user-space stp
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Nikolay Aleksandrov <nikolay@cumulusnetworks.com>
commit 7ce42de1895d4787b47b004638d642dcacb464fe

When STP is running in user-space and querier is configured, the
querier timer is not started when a port goes to a non-blocking state.
This patch unifies the user- and kernel-space stp multicast port enable
path and enables it in all states different from blocking. Note that when a
port goes in BR_STATE_DISABLED it's not enabled because that is handled
in the beginning of the port list loop.

	Signed-off-by: Nikolay Aleksandrov <nikolay@cumulusnetworks.com>
	Acked-by: Herbert Xu <herbert@gondor.apana.org.au>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 7ce42de1895d4787b47b004638d642dcacb464fe)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>