sctp: keep fragmentation point aligned to word size
jira LE-1907
cve CVE-2016-9555
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Marcelo Ricardo Leitner <marcelo.leitner@gmail.com>
commit 659e0bcaebc4ca36e64eac6e9f39c1904b17472c

If the user supply a different fragmentation point or if there is a
network header that cause it to not be aligned, force it to be aligned.

Fragmentation point at a value that is not aligned is not optimal.  It
causes extra padding to be used and has just no pros.

v2:
 - Make use of the new WORD_TRUNC macro

	Signed-off-by: Marcelo Ricardo Leitner <marcelo.leitner@gmail.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 659e0bcaebc4ca36e64eac6e9f39c1904b17472c)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>