fm10k: Cleanup exception handling for mailbox interrupt
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Alexander Duyck <aduyck@mirantis.com>
commit e00e23bceba48a8f0c94fefe26948404cbd43d0a

This patch addresses two issues.

First is the fact that the fm10k_mbx_free_irq was assuming msix_entries was
valid and that will not always be the case.  As such we need to add a check
for if it is NULL.

Second is the fact that we weren't freeing the IRQ if the mailbox API
returned an error on trying to connect.

	Signed-off-by: Alexander Duyck <aduyck@mirantis.com>
	Reviewed-by: Bruce Allan <bruce.w.allan@intel.com>
	Tested-by: Krishneil Singh <Krishneil.k.singh@intel.com>
	Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit e00e23bceba48a8f0c94fefe26948404cbd43d0a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>