fm10k: Fix multicast mode sync issues
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Ngai-Mint Kwan <ngai-mint.kwan@intel.com>
commit 8998763a7b57583ef2e07f68ea6a7d05bcfc1cfa

Multicast mode checking is no longer a requirement to perform unicast
and multicast address syncs. Specifically, a device operating in
promiscuous and/or all multicast mode is not excluded. The issue occurs
when the netdev is pre-configured to either multicast mode and is
enabled for the first time. The multicast-group table in the Switch
Manager will be missing obvious multicast entries associated to this
netdev.

Changes were also made to disallow unicast and multicast syncs with
VLAN 0. The Switch Manager considers VLAN 0 to be an invalid entry.
Requests with VLAN 0 by the netdev are only generated when the driver is
freshly installed and the default VID is not set.

	Signed-off-by: Ngai-Mint Kwan <ngai-mint.kwan@intel.com>
	Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
	Tested-by: Krishneil Singh <Krishneil.k.singh@intel.com>
	Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 8998763a7b57583ef2e07f68ea6a7d05bcfc1cfa)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>