dell-laptop: Only install the i8042 filter when rfkill is active
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Hans de Goede <hdegoede@redhat.com>
commit 97f440c23f2b02fac8af0558cba9ca0bed603794

Installing the i8042 filter is not useful on machines where rfkill is not
whitelisted, so move the filter installation into dell_setup_rfkill,
after the whitelist check.

This avoids doing a needless and potentially troublesome rfkill query
(dell_send_request(buf, 17, 11)) when the wireless Fn key gets pressed on
non whitelisted laptops.

This patch was written as a result of:
https://bugzilla.redhat.com/show_bug.cgi?id=1045807
It is not yet clear if this is related, but it is a good idea to not register
the i8042 filter in general.

	Signed-off-by: Hans de Goede <hdegoede@redhat.com>
	Signed-off-by: Matthew Garrett <matthew.garrett@nebula.com>
(cherry picked from commit 97f440c23f2b02fac8af0558cba9ca0bed603794)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>