vfio: Enable VFIO device for powerpc
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author David Gibson <david@gibson.dropbear.id.au>
commit 178a787502123b01499c5a4617b94bb69ad49dd5

ec53500f "kvm: Add VFIO device" added a special KVM pseudo-device which is
used to handle any necessary interactions between KVM and VFIO.

Currently that device is built on x86 and ARM, but not powerpc, although
powerpc does support both KVM and VFIO.  This makes things awkward in
userspace

Currently qemu prints an alarming error message if you attempt to use VFIO
and it can't initialize the KVM VFIO device.  We don't want to remove the
warning, because lack of the KVM VFIO device could mean coherency problems
on x86.  On powerpc, however, the error is harmless but looks disturbing,
and a test based on host architecture in qemu would be ugly, and break if
we do need the KVM VFIO device for something important in future.

There's nothing preventing the KVM VFIO device from being built for
powerpc, so this patch turns it on.  It won't actually do anything, since
we don't define any of the arch_*() hooks, but it will make qemu happy and
we can extend it in future if we need to.

	Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
	Reviewed-by: Eric Auger <eric.auger@linaro.org>
	Signed-off-by: Paul Mackerras <paulus@samba.org>
(cherry picked from commit 178a787502123b01499c5a4617b94bb69ad49dd5)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>