qlge: fix a timeout loop in ql_change_rx_buffers()
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Dan Carpenter <dan.carpenter@oracle.com>
commit 351434c6ba92e1fe7799a0c33c1412584a0fb3de

The problem here is that after the loop we test for "if (!i) " but
because "i--" is a post-op we exit with i set to -1.  I have fixed this
by changing it to a pre-op instead.  I had to change the starting value
from 3 to 4 so that we still iterate 3 times.

	Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 351434c6ba92e1fe7799a0c33c1412584a0fb3de)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>