balloon: check the number of available pages in leak balloon
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Konstantin Neumoin <kneumoin@virtuozzo.com>
commit 37cf99e08c6fb4dcea0f9ad2b13b6daa8c76a711

The balloon has a special mechanism that is subscribed to the oom
notification which leads to deflation for a fixed number of pages.
The number is always fixed even when the balloon is fully deflated.
But leak_balloon did not expect that the pages to deflate will be more
than taken, and raise a "BUG" in balloon_page_dequeue when page list
will be empty.

So, the simplest solution would be to check that the number of releases
pages is less or equal to the number taken pages.

	Cc: stable@vger.kernel.org
	Signed-off-by: Konstantin Neumoin <kneumoin@virtuozzo.com>
	Signed-off-by: Denis V. Lunev <den@openvz.org>
CC: Michael S. Tsirkin <mst@redhat.com>
	Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
(cherry picked from commit 37cf99e08c6fb4dcea0f9ad2b13b6daa8c76a711)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>