vfio iommu type1: Fix size argument to vfio_find_dma() in pin_pages/unpin_pages
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Kirti Wankhede <kwankhede@nvidia.com>
commit 2b8bb1d771f736b8b34bf160115aee1b12d29f83

Passing zero for the size to vfio_find_dma() isn't compatible with
matching the start address of an existing vfio_dma. Doing so triggers a
corner case. In vfio_find_dma(), when the start address is equal to
dma->iova and size is 0, check for the end of search range makes it to
take wrong side of RB-tree. That fails the search even though the address
is present in mapped dma ranges.
In functions pin_pages and unpin_pages, the iova which is being searched
is base address of page to be pinned or unpinned. So here size should be
set to PAGE_SIZE, as argument to vfio_find_dma().

	Signed-off-by: Kirti Wankhede <kwankhede@nvidia.com>
	Signed-off-by: Neo Jia <cjia@nvidia.com>
	Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
(cherry picked from commit 2b8bb1d771f736b8b34bf160115aee1b12d29f83)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>