raid1: handle read error also in readonly mode
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Tomasz Majchrzak <tomasz.majchrzak@intel.com>
commit 7449f699b2fb23bdee0a0f03aa4efb5f96fd403f
Empty-Commit: Cherry-Pick Conflicts during history rebuild.
Will be included in final tarball splat. Ref for failed cherry-pick at:
ciq/ciq_backports/kernel-3.10.0-693.el7/7449f699.failed

If write is the first operation on a disk and it happens not to be
aligned to page size, block layer sends read request first. If read
operation fails, the disk is set as failed as no attempt to fix the
error is made because array is in auto-readonly mode. Similarily, the
disk is set as failed for read-only array.

Take the same approach as in raid10. Don't fail the disk if array is in
readonly or auto-readonly mode. Try to redirect the request first and if
unsuccessful, return a read error.

	Signed-off-by: Tomasz Majchrzak <tomasz.majchrzak@intel.com>
	Signed-off-by: Shaohua Li <shli@fb.com>
(cherry picked from commit 7449f699b2fb23bdee0a0f03aa4efb5f96fd403f)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>

# Conflicts:
#	drivers/md/raid1.c