openvswitch: fix flow stats accounting when node 0 is not possible
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Thadeu Lima de Souza Cascardo <cascardo@redhat.com>
commit 40773966ccf1985a1b2bb570a03cbeaf1cbd4e00

On a system with only node 1 as possible, all statistics is going to be
accounted on node 0 as it will have a single writer.

However, when getting and clearing the statistics, node 0 is not going
to be considered, as it's not a possible node.

Tested that statistics are not zero on a system with only node 1
possible. Also compile-tested with CONFIG_NUMA off.

	Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@redhat.com>
	Acked-by: Pravin B Shelar <pshelar@ovn.org>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 40773966ccf1985a1b2bb570a03cbeaf1cbd4e00)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>