mpt3sas: Ensure the connector_name string is NUL-terminated
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Calvin Owens <calvinowens@fb.com>
commit 310c8e40d5cd2512b6d46177563460726cc6166a

We blindly trust the hardware to give us NUL-terminated strings, which
is a bad idea because it doesn't always do that. For example:

  [  481.184784] mpt3sas_cm0: 	enclosure level(0x0000), connector name(     \x3)

In this case, connector_name is four spaces. We got lucky here because
the 2nd byte beyond our character array happens to be a NUL. Fix this by
explicitly writing '\0' to the end of the string to ensure we don't run
off the edge of the world in printk().

	Signed-off-by: Calvin Owens <calvinowens@fb.com>
	Acked-by: Chaitra P B <chaitra.basappa@broadcom.com>
	Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 310c8e40d5cd2512b6d46177563460726cc6166a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>